Move writer to sub directories.
